+++ src/app/modules/Screens/LoginPage/LoginPage.js
@@ -6,7 +6,7 @@ import Card from "@material-ui/core/Card";
 
 import styled from "styled-components";
 
-import Button from "../../components/Button";
+import Button from "../../../components/Button";
 
 const Container = styled.div`
   display: flex;
@@ -14,8 +14,7 @@ const Container = styled.div`
   justify-content: flex-start;
   align-items: center;
   background-color: #89cff0;
-  height: auto;
-  min-height: 100vh;
+  height: 100vh;
   width: 100%;
 `;
 
@@ -25,9 +24,9 @@ const Logo = styled.div`
   justify-content: center;
   align-items: center;
   width: 100%;
-  flex: 1;
+  flex: 2;
   font-family: "Bungee", cursive;
-  font-size: 4rem;
+  font-size: 7rem;
 `;
 
 const FormSection = styled.div`
@@ -43,72 +42,50 @@ const FormSection = styled.div`
   }
 `;
 
-const TermsLinks = styled.div`
+const RegisterLinks = styled.div`
   display: flex;
   flex-direction: column;
   justify-content: center;
   align-items: center;
   font-size: 1.4rem;
-  text-align: center;
-  margin-bottom: 2rem;
-
-  a {
-    margin-top: 1rem;
-  }
 `;
 
 const FormContainer = styled(Card)`
   display: flex;
   flex-direction: column;
-  justify-content: flex-start;
+  justify-content: center;
   align-items: center;
   width: 30rem;
-  padding: 2rem 3rem;
+  height: 30rem;
+  padding: 2rem;
 `;
 
 const Form = styled.form`
   display: flex;
   flex-direction: column;
-  justify-content: flex-start;
+  justify-content: space-around;
   align-items: center;
   width: 100%;
-  margin-bottom: 2rem;
+  height: 100%;
+  padding: 2rem 3rem;
 `;
 
 const StyledInput = styled(Input)`
   input {
     font-size: 1.6rem;
-    padding-top: 2.5rem;
   }
 `;
 
-class RegisterPage extends Component {
-  state = {
-    userName: "",
-    wages: "",
-    email: "",
-    password: "",
-    confirmPassword: ""
-  };
-
+class LoginPage extends Component {
   submitForm(e) {
     e.preventDefault();
-    const { userName, wages, email, password, confirmPassword } = this.state;
-    if (
-      userName &&
-      wages &&
-      email &&
-      password &&
-      confirmPassword &&
-      password === confirmPassword
-    ) {
-      this.props.registerNewAccount(this.state);
-    }
+    const { email, password } = this.props.auth;
+    this.props.logInWithEmailAndPassword(email, password);
   }
 
   handleInputChange(field, event) {
     const { value } = event.target;
-    this.setState({ [field]: value });
+    this.props.updateInput(field, value);
   }
 
   render() {
@@ -116,7 +93,7 @@ class RegisterPage extends Component {
       return <Redirect to="/today" />;
     }
 
-    const { userName, wages, email, password, confirmPassword } = this.state;
+    const { email, password } = this.props;
     return (
       <Container>
         <Logo>
@@ -127,28 +104,11 @@ class RegisterPage extends Component {
           <FormContainer>
             <Form onSubmit={event => this.submitForm(event)}>
               <StyledInput
-                type="string"
-                value={userName}
-                onChange={e => this.handleInputChange("userName", e)}
-                placeholder="user name"
-                fullWidth
-                required
-              />
-              <StyledInput
-                type="number"
-                value={wages}
-                onChange={e => this.handleInputChange("wages", e)}
-                placeholder="current hourly salary"
-                fullWidth
-                required
-              />
-              <StyledInput
                 type="email"
                 value={email}
                 onChange={e => this.handleInputChange("email", e)}
                 placeholder="email"
                 fullWidth
-                required
               />
               <StyledInput
                 type="password"
@@ -156,41 +116,22 @@ class RegisterPage extends Component {
                 onChange={e => this.handleInputChange("password", e)}
                 placeholder="password"
                 fullWidth
-                required
-              />
-              <StyledInput
-                type="password"
-                value={confirmPassword}
-                onChange={e => this.handleInputChange("confirmPassword", e)}
-                placeholder="confirm password"
-                fullWidth
-                required
               />
               <Button
                 type="submit"
                 color="primary"
                 variant="contained"
                 fullWidth
-                style={{ marginTop: "2rem" }}
               >
-                Register
+                Login
               </Button>
             </Form>
-            <TermsLinks>
-              <span>
-                By registering you automatically accept the following:
-              </span>
+            <RegisterLinks>
+              <span>Don't have an account?</span>
               <Link to="/register">
-                <span>Terms and Conditions</span>
+                <span>Register here.</span>
               </Link>
-              <Link to="/register">
-                <span>Privacy policy</span>
-              </Link>
-            </TermsLinks>
-            <TermsLinks>
-              Already have an account?
-              <Link to="/">Login</Link>
-            </TermsLinks>
+            </RegisterLinks>
           </FormContainer>
         </FormSection>
       </Container>
@@ -198,4 +139,4 @@ class RegisterPage extends Component {
   }
 }
 
-export default RegisterPage;
+export default LoginPage;
