+++ src/app/modules/Screens/RegisterPage/RegisterPage.js
@@ -6,7 +6,7 @@ import Card from "@material-ui/core/Card";
 
 import styled from "styled-components";
 
-import Button from "../../components/Button";
+import Button from "../../../components/Button";
 
 const Container = styled.div`
   display: flex;
@@ -14,7 +14,8 @@ const Container = styled.div`
   justify-content: flex-start;
   align-items: center;
   background-color: #89cff0;
-  height: 100vh;
+  height: auto;
+  min-height: 100vh;
   width: 100%;
 `;
 
@@ -24,9 +25,9 @@ const Logo = styled.div`
   justify-content: center;
   align-items: center;
   width: 100%;
-  flex: 2;
+  flex: 1;
   font-family: "Bungee", cursive;
-  font-size: 7rem;
+  font-size: 4rem;
 `;
 
 const FormSection = styled.div`
@@ -42,50 +43,72 @@ const FormSection = styled.div`
   }
 `;
 
-const RegisterLinks = styled.div`
+const TermsLinks = styled.div`
   display: flex;
   flex-direction: column;
   justify-content: center;
   align-items: center;
   font-size: 1.4rem;
+  text-align: center;
+  margin-bottom: 2rem;
+
+  a {
+    margin-top: 1rem;
+  }
 `;
 
 const FormContainer = styled(Card)`
   display: flex;
   flex-direction: column;
-  justify-content: center;
+  justify-content: flex-start;
   align-items: center;
   width: 30rem;
-  height: 30rem;
-  padding: 2rem;
+  padding: 2rem 3rem;
 `;
 
 const Form = styled.form`
   display: flex;
   flex-direction: column;
-  justify-content: space-around;
+  justify-content: flex-start;
   align-items: center;
   width: 100%;
-  height: 100%;
-  padding: 2rem 3rem;
+  margin-bottom: 2rem;
 `;
 
 const StyledInput = styled(Input)`
   input {
     font-size: 1.6rem;
+    padding-top: 2.5rem;
   }
 `;
 
-class LoginPage extends Component {
+class RegisterPage extends Component {
+  state = {
+    userName: "",
+    wages: "",
+    email: "",
+    password: "",
+    confirmPassword: ""
+  };
+
   submitForm(e) {
     e.preventDefault();
-    const { email, password } = this.props.auth;
-    this.props.logInWithEmailAndPassword(email, password);
+    const { userName, wages, email, password, confirmPassword } = this.state;
+    if (
+      userName &&
+      wages &&
+      email &&
+      password &&
+      confirmPassword &&
+      password === confirmPassword
+    ) {
+      this.props.registerNewAccount(this.state);
+    }
   }
 
   handleInputChange(field, event) {
     const { value } = event.target;
-    this.props.updateInput(field, value);
+    this.setState({ [field]: value });
   }
 
   render() {
@@ -93,7 +116,7 @@ class LoginPage extends Component {
       return <Redirect to="/today" />;
     }
 
-    const { email, password } = this.props;
+    const { userName, wages, email, password, confirmPassword } = this.state;
     return (
       <Container>
         <Logo>
@@ -104,11 +127,28 @@ class LoginPage extends Component {
           <FormContainer>
             <Form onSubmit={event => this.submitForm(event)}>
               <StyledInput
+                type="string"
+                value={userName}
+                onChange={e => this.handleInputChange("userName", e)}
+                placeholder="user name"
+                fullWidth
+                required
+              />
+              <StyledInput
+                type="number"
+                value={wages}
+                onChange={e => this.handleInputChange("wages", e)}
+                placeholder="current hourly salary"
+                fullWidth
+                required
+              />
+              <StyledInput
                 type="email"
                 value={email}
                 onChange={e => this.handleInputChange("email", e)}
                 placeholder="email"
                 fullWidth
+                required
               />
               <StyledInput
                 type="password"
@@ -116,22 +156,41 @@ class LoginPage extends Component {
                 onChange={e => this.handleInputChange("password", e)}
                 placeholder="password"
                 fullWidth
+                required
+              />
+              <StyledInput
+                type="password"
+                value={confirmPassword}
+                onChange={e => this.handleInputChange("confirmPassword", e)}
+                placeholder="confirm password"
+                fullWidth
+                required
               />
               <Button
                 type="submit"
                 color="primary"
                 variant="contained"
                 fullWidth
+                style={{ marginTop: "2rem" }}
               >
-                Login
+                Register
               </Button>
             </Form>
-            <RegisterLinks>
-              <span>Don't have an account?</span>
+            <TermsLinks>
+              <span>
+                By registering you automatically accept the following:
+              </span>
               <Link to="/register">
-                <span>Register here.</span>
+                <span>Terms and Conditions</span>
               </Link>
-            </RegisterLinks>
+              <Link to="/register">
+                <span>Privacy policy</span>
+              </Link>
+            </TermsLinks>
+            <TermsLinks>
+              Already have an account?
+              <Link to="/">Login</Link>
+            </TermsLinks>
           </FormContainer>
         </FormSection>
       </Container>
@@ -139,4 +198,4 @@ class LoginPage extends Component {
   }
 }
 
-export default LoginPage;
+export default RegisterPage;
